Add MSTORE8 opcode spec
