+++ src/zkevm_specs/opcode/__init__.py
@@ -3,3 +3,4 @@ from .byte import *
 from .comparator import *
 from .lt_gt import *
 from .stack import *
+from .mload_mstore import *
